requirements.txt
pandas>=1.3
pyarrow
jsonschema
lxml
tqdm
pytest
folium
geopandas
sqlalchemy
psycopg2-binary
pyspark